gnutls: Update to 3.7.4
Bump to latest upstream release.  Rebased existing patches.

This version introduced certificate compression support, and with that,
an optional dependency on zstd. When zstd is built, configure will pick
it up, and build will fail due to a missing dependency:

Package libgnutls is missing dependencies for the following libraries:
libzstd.so.1

Disable zstd support to avoid that.

Signed-off-by: John Audia <graysky@archlinux.us>
[disable zstd]
Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>