node-*: support for npm@8 (#414)
* node-homebridge: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

* Revert "node-yarn: bump to v1.22.17"

This reverts commit 38b18a50f22ff132596439adca0cdef98bedbb9a.

* node-yarn: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

* node-hid: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

* javascript-obfuscator: bump to 2.19.0 and switch to autorelease

New version of package node-javascript-obfuscator.

Signed-off-by: Zbyněk Kocur <zbynek.kocur@fel.cvut.cz>

* node-javascript-obfuscator: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

The modification method is different from other node modules.
The reason is due to the npm@8 issue.
https://github.com/npm/cli/issues/4027

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

* node-serialport: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

* node-cylon: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

* node-arduino-firmata: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

* node-serialport-bindings: Support for npm@8

With the upgrade of node.js to version 16, the npm version will also change to version 8.
This fix is to support npm@8. npm@6 can also build without problems.

Signed-off-by: Hirokazu MORIKAWA <morikw2@gmail.com>

Co-authored-by: Hirokazu MORIKAWA <morikw2@gmail.com>
Co-authored-by: Zbyněk Kocur <zbynek.kocur@fel.cvut.cz>