iputils: provide iputils-ping6 in iputils-ping
iputils-ping6 was a subpackage of the iputils package providing the
ping4 and ping6 command before iputils was moved from core to packages.
Currently ping4 and ping6 are replaced by ping -4/-6 and compatibility
symlinks are only installed when explicitly told so with an option, but
the functionality is always provided by iputils-ping.

Signed-off-by: Michal Vasilek <michal.vasilek@nic.cz>