pulseaudio: simplify and rework Makefile
There were two conffiles sections and both of them were same, but for
different variants. We can have just one conffile section and use it also
for the other variant.

The same applies for the install section for different variants.
- We have two install sections, but we call the first one with the same
  files and then add something more for the second variant.

- While at it to make it easier, let's change those three rows for
  copying packages into the single one to make sure that I did not miss
anything. Also, we create a directory first and then move files.

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>