ices: Update to 2.0.3
This fixes a timing issue causing a playlist to be streamed as-fast-as-
possible until memory exhaustion, finally crashing ices and making it
basically useless.

This problem occurred in  the timing code in src/common/timing/ and is
probably related to libc header changes because ices 2.0.2 worked up until
OpenWrt 21.02.3 (at least) but failed in OpenWrt 22.03.2.

Signed-off-by: Tilman Vogel <tilman.vogel@web.de>