libwebsockets: fix build with gcc 13 (#760)
with gcc 13 libwebsockets fails with -Werror=enum-int-mismatch with both
SSL variants. These 2 patches work out that issue. Problem is mentioned
as a issue in openwrt/packages#20949 and instead of overriding issue
with appropriate CFLAGS, this approach was suggested for me as it's
been made as a PR (for openssl) at warmcat/libwebsockets#2824

Second patch for mbedtls, was made by me using same approach
as was used of openssl.

Eventually these propably will be merged into libwebsockets mainstream.

Signed-off-by: Oskari Rauta <oskari.rauta@gmail.com>
Co-authored-by: Oskari Rauta <oskari.rauta@gmail.com>